// src/common/enums/domain-exception-code.enum.ts
export enum DomainExceptionCode {
  ENTITY_NOT_FOUND = 'ENTITY_NOT_FOUND',
  ENTITY_ALREADY_EXISTS = 'ENTITY_ALREADY_EXISTS',
  SLUG_ALREADY_EXISTS = 'SLUG_ALREADY_EXISTS',
  DATABASE_ERROR = 'DATABASE_ERROR',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_RESET_TOKEN = 'INVALID_RESET_TOKEN',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  FORBIDDEN_OPERATION = 'FORBIDDEN_OPERATION',
  INVITE_ALREADY_EXISTS = 'INVITE_ALREADY_EXISTS',
  INVITE_EXPIRED = 'INVITE_EXPIRED',
  INSUFFICIENT_SUBSCRIPTION = 'INSUFFICIENT_SUBSCRIPTION',
  ACTIVE_SUBSCRIPTION_REQUIRED = 'ACTIVE_SUBSCRIPTION_REQUIRED',
  CANNOT_DOWNGRADE = 'CANNOT_DOWNGRADE',
  ASAAS_API_ERROR = 'ASAAS_API_ERROR',
}